mas phpstan

+++ src/cambios/db/DBEsquema.php
@@ -641,6 +641,10 @@ class DBEsquema extends DBAbstract
         $filename = $datosTabla['filename'];
         $oDbl = $this->oDbl;
 
+        if ($oDbl === null) {
+            throw new \RuntimeException(_('No hay conexión a la base de datos.'));
+        }
+
         $a_sql = [];
         $a_sql[0] = "TRUNCATE $nom_tabla RESTART IDENTITY CASCADE;";
         $this->executeSql($a_sql);
@@ -689,6 +693,10 @@ class DBEsquema extends DBAbstract
         $id_seq = $datosTabla['id_seq'];
         $filename = $datosTabla['filename'];
         $oDbl = $this->oDbl;
+
+        if ($oDbl === null) {
+            throw new \RuntimeException(_('No hay conexión a la base de datos.'));
+        }
 
         /* Se borra con la primera.
         $a_sql = [];
